% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cetaceans}
\alias{cetaceans}
\title{Ages of stranded dolphins}
\format{A data frame with 30 rows and 2 variables:
\describe{
  \item{age}{age of the dead animals}
  \item{M}{number of animals dead at the corresponding age}
  ...
}}
\usage{
cetaceans
}
\description{
A dataset containing the age-at-death of 435 stranded dolphins.
}
\keyword{datasets}
