% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.Leslie.R
\name{cohort}
\alias{cohort}
\title{Cohort method}
\usage{
cohort(width12)
}
\arguments{
\item{width12}{Width of the first two age classes.}
}
\value{
A two element vector representing the first two values of the nax column of the life table.
}
\description{
Dummy function called in \code{\link{life.Leslie}}
}
\note{
Utility function called by \code{\link{life.Leslie}}.
}
\seealso{
\code{\link{life.Leslie}}
}
\keyword{cohort}
\keyword{leslie-matrix}
\keyword{life-table}
