% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.Leslie.R
\name{keyfitz}
\alias{keyfitz}
\title{Keyfitz and Flieger method.}
\usage{
keyfitz(b0, nMx)
}
\arguments{
\item{b0}{Two-element vector of regression coefficients provided in Keyfitz and Flieger (1990).}

\item{nMx}{Period central death rates = nDx/nKx.}
}
\value{
The first two values (age classes 0-1 and 1-5) of the nax column of a period life table.
}
\description{
Utility used by \code{\link{life.Leslie}} to graduate the person-years lived by those dying in the interval by the
method of Keyfitz and Flieger (1990).
}
\note{
Utility function called by \code{\link{life.Leslie}}
}
\references{
Keyfitz, N. and Flieger, W. (1990). World population growth and aging: Demographic trends in the late twentieth century. Chicago: University of Chicago Press.
}
\seealso{
\code{\link{life.Leslie}}.
}
\keyword{keyfitz}
\keyword{leslie-matrix}
\keyword{life-table}
