% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HP.mod.R
\name{ll.binom}
\alias{ll.binom}
\title{Binomial likelihood.}
\usage{
ll.binom(x, n, p)
}
\arguments{
\item{x}{Same as x from \link{dbinom} - successes.}

\item{n}{Same as size from \link{dbinom} - trials.}

\item{p}{Same as prob from \link{dbinom} - observed probability.}
}
\description{
Calculates a log likelihood from a binomial distribution.
}
\references{
R Development Core Team. (2009). R: A Language and Environment for Statistical Computing. R Foundation for Statistical Computing Vienna, Austria.
}
\seealso{
\link{dbinom}
}
\keyword{Heligman-Pollard}
\keyword{binomial}
\keyword{bycatch}
\keyword{mortality}
