% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HP.mod.R
\name{mod}
\alias{mod}
\title{Heligman-Pollard parameter coversion to age-specific probabilites of death.}
\usage{
mod(theta, x)
}
\arguments{
\item{theta}{A vector containing values for the 9 parameters of the adapted Heligman-Pollard model.}

\item{x}{A vector containing the ages at which to calculate the probabilities of death.}
}
\value{
A vector of probabilities of death at ages defined by x.
}
\description{
Calculates the age-specific probabilities of death using the Heligman-Pollard model.
}
\references{
Heligman, L. and Pollard, J.H. (1980). The Age Pattern of Mortality. Journal of the Institute of Actuaries 107:49–80.
}
\keyword{Heligman-Pollard}
\keyword{bycatch}
\keyword{mortality}
