% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HP.CI.R
\name{mod.risk}
\alias{mod.risk}
\title{Heligman-Pollard parameter coversion to age-specific probabilites of death due to an external risk.}
\usage{
mod.risk(theta, x)
}
\arguments{
\item{theta}{A vector containing values for the 9 parameters of the adapted Heligman-Pollard model.}

\item{x}{A vector containing the ages at which to calculate the probabilities of death.}
}
\value{
A vector of probabilities of death at ages due to an external risk.
}
\description{
Calculates the age-specific probabilities of death due to an external risk using the Heligman-Pollard model.
}
\note{
Utility function called by \code{\link{HP.CI}}.
}
\references{
Heligman, L. and Pollard, J.H. (1980). The Age Pattern of Mortality. Journal of the Institute of Actuaries 107:49–80.
}
\seealso{
\code{\link{HP.CI}}
}
\keyword{Heligman-Pollard}
\keyword{bycatch}
\keyword{mortality}
